Everywhere: Change Display Method to CollapsingHeaders
Information is grouped into collapsing headers which allows for more
screen space for graphs and for cleaner segmenting of information. Each
collapsing header has its own table to format the information within it
into columns and rows.

Eventually, I would like to have to CollapsingHeaders each have a graph
that goes with them that is displayed above all the CollapsingHeaders
that when a CollapsingHeader is clicked the graph for that information
segment appears at the top.