Everywhere: Remove Dear Imgui submodule and use only the source files
Removed Dear Imgui submodule becauase it was a mess in terms of the
azure pipeline. Pulled in all the newest imgui files but just added
them to a file as part of the SystemMonitor repository and not as a
submodule. Much easier to work with and guarantees stability. Future
updates are a little more cumbersome to us but no big deal, just copy
files from git repo and add them manually instead of cloning project.