Formatting: Add function to place text on at specific location
insert_text_at(Window_Width_Multiple, std::string const) takes a
a multiple and the text to be inserted. The cursor position is
calculated based on the multiple. This way it makes it easier for the
user to place and align the cursor.

The window width is multiplied by the multiple and then the text_width
is divided by 2 and subtracted from the window width * multiple. What
this does is first, get the rough location of where the user wants to
place their text, then divide the text in half and subtract so that the
placement is not shifted by the length of the text.